refactor(event-governance): author rule packs under agent/event-rules
Move the event-rules asset tree (schema, packs, parity fixtures) from
contracts/ to agent/event-rules/, alongside agent/workflows/ — one asset
tree, one authoring UX, and the natural source for the future
RegistryService (#246).

The agent runtime never loads these files (it receives resolved
event_rules on the TaskRecord), so only the three file consumers move:
the CDK pack resolver (bundle-time import), the CLI --fixture loader, and
the Python + TS parity/evaluator tests. Dual-plane parity is unchanged —
the CDK-resolved rules remain the single source both planes consume.

Repoints resolver import, CLI fixture path, four test fixture paths,
blueprint schema comment, and EVENT_GOVERNANCE.md (+ Starlight mirror).
