+++ agent/event-rules-parity/README.md
@@ -5,7 +5,7 @@ Golden-file vectors shared by:
 - `agent/tests/test_event_rules_parity.py` (Python evaluator)
 - `cdk/test/handlers/shared/event-rules-parity.test.ts` (TypeScript evaluator)
 
-Fixtures live under `contracts/event-rules/fixtures/` and are referenced by
+Fixtures live under `agent/event-rules/fixtures/` and are referenced by
 basename from this parity contract.
 
 Design reference: `docs/design/EVENT_GOVERNANCE.md` (issue #230).
