+++ agent/event-rules/fixtures/observe-repo-setup.json
