+++ agent/event-rules/packs/platform-default-v1.json
